convert to epoch 112

+++ global.h
@@ -99,6 +99,6 @@ char* g_pool_unitOfMeasurement = nullptr;
 int64_t g_pool_numberOfShares = -1;
 char g_pool_numberOfDecimalPlaces = 0;
 
-int64_t g_qpool_get_infor_number_of_pool = 0;
+int32_t g_qpool_get_infor_number_of_pool = 0;
 
 uint64_t g_get_log_passcode[4] = {0};
